move pdata up one
