+++ rust/otel-arrow-rust/src/lib.rs
@@ -22,8 +22,8 @@ mod schema;
 #[cfg(test)]
 mod test_util;
 
-// Add the proto module which exposes the generated code and our custom types
 pub mod proto;
+pub mod pdata;
 
 #[path = ""]
 pub mod opentelemetry {
