+++ rust/otel-arrow-rust/src/proto/mod.rs
@@ -10,8 +10,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod pdata;
-
 #[path = "."]
 pub mod opentelemetry {
 	#[path = "."]
