+++ rust/otel-arrow-rust/src/proto/opentelemetry.proto.experimental.arrow.v1.rs
@@ -1,7 +1,7 @@
 // This file is @generated by prost-build.
 /// A message sent by an exporter to a collector containing a batch of Arrow
 /// records.
-#[derive(crate::proto::pdata::otap::Message)]
+#[derive(crate::pdata::otap::Message)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct BatchArrowRecords {
     /// \[mandatory\] Batch ID. Must be unique in the context of the stream.
@@ -15,7 +15,7 @@ pub struct BatchArrowRecords {
     pub headers: ::prost::alloc::vec::Vec<u8>,
 }
 /// Represents a batch of OTel Arrow entities.
-#[derive(crate::proto::pdata::otap::Message)]
+#[derive(crate::pdata::otap::Message)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct ArrowPayload {
     /// \[mandatory\] A canonical ID representing the schema of the Arrow Record.
@@ -37,7 +37,7 @@ pub struct ArrowPayload {
     pub record: ::prost::alloc::vec::Vec<u8>,
 }
 /// A message sent by a Collector to the exporter that opened the data stream.
-#[derive(crate::proto::pdata::otap::Message)]
+#[derive(crate::pdata::otap::Message)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct BatchStatus {
     #[prost(int64, tag = "1")]
@@ -49,7 +49,7 @@ pub struct BatchStatus {
 }
 /// Enumeration of all the OTel Arrow payload types currently supported by the
 /// OTel Arrow protocol.
-#[derive(crate::proto::pdata::otap::Message)]
+#[derive(crate::pdata::otap::Message)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum ArrowPayloadType {
@@ -161,7 +161,7 @@ impl ArrowPayloadType {
 }
 /// StatusCode carries certain known meanings in Arrow.  Values match
 /// the gRPC code space, see <https://grpc.io/docs/guides/status-codes/.>
-#[derive(crate::proto::pdata::otap::Message)]
+#[derive(crate::pdata::otap::Message)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, ::prost::Enumeration)]
 #[repr(i32)]
 pub enum StatusCode {
