+++ rust/otel-arrow-rust/src/proto/opentelemetry.proto.resource.v1.rs
@@ -1,6 +1,6 @@
 // This file is @generated by prost-build.
 /// Resource information.
-#[derive(crate::proto::pdata::otlp::Message)]
+#[derive(crate::pdata::otlp::Message)]
 #[derive(Clone, PartialEq, ::prost::Message)]
 pub struct Resource {
     /// Set of attributes that describe the resource.
